docs Update to README
docs: Update this document so it has all of the necessary information for a README file and use a formatting style guide

"Before, this README file had nothing in it.  Now, it has something in it and it follows a style guide.  This README file will need more work done to it in order to be anywhere complete."